Removing overly restrictive version cap on Microsoft.Extensions packages
Version 6.0.0 of the Microsoft.Extensions.** packages all support `netstandard2.0` so can be used all the way back to `netcoreapp2.1`.

This is causing issues where some apps that are still on `netcoreapp3.1` are using another package that requires Microsoft.Extensions.Logging.Console 5.0.0+ which references Microsoft.Extensions.DependencyInjection.Abstractions 5.0.0. But this package has a requirement of <4.0.0 for M.E.DI.A (despite the fact that there is nothing otherwise that would prevent 5.0.0+ from being used).

Note that this won't change the version used for any existing usages. It will only allow other use cases to opt in to reference a higher version.